// src/dora/enums/companies.enum.ts
/**
 * 32 3PSPs.
 */
export enum ECompanies {
  AFAS = 'AFAS',
  BITVAVO = 'BITVAVO',
  BITWARDEN = 'BITWARDEN',
  BLOCKRISE = 'BLOCKRISE',
  BUNQ = 'BUNQ',
  CHAINSTACK = 'CHAINSTACK',
  CLOUDFLARE = 'CLOUDFLARE',
  COLT = 'COLT',
  COINGECKO = 'COINGECKO',
  COMPLYADVANTAGE = 'COMPLYADVANTAGE',
  ELLIPTIC = 'ELLIPTIC',
  GITHUB = 'GITHUB',
  GOOGLE = 'GOOGLE',
  IDIN = 'IDIN',
  KRAKEN = 'KRAKEN',
  LINEAR = 'LINEAR',
  MICROSOFT = 'MICROSOFT',
  NOTION = 'NOTION',
  POSTCODEAPI = 'POSTCODEAPI',
  PAY_DOT = 'PAY_DOT',
  SANITY = 'SANITY',
  SCRIVE = 'SCRIVE',
  SECUROSYS = 'SECUROSYS',
  SENDGRID = 'SENDGRID',
  SENTRY = 'SENTRY',
  SLACK = 'SLACK',
  SONARCLOUD = 'SONARCLOUD',
  SUREPAY = 'SUREPAY',
  TWILIO = 'TWILIO',
  UNIFI = 'UNIFI',
  VERCEL = 'VERCEL',
  VERIFF = 'VERIFF',
}